+++ application/views/fleetdetail.php
@@ -1,0 +1,38 @@
+<div class="row">
+    <div class="card-body">
+        <div class="table-responsive">
+            <table class="table table-default" id="dataTable" width="100%" cellspacing="0">
+                    <thead>
+                    <tr>
+                        <th>ID</th>
+                        <th>Plane Name</th>
+                        <th>Manufacturer</th>
+                        <th>Model</th>
+                        <th>Price</th>
+                        <th>Seats</th>
+                        <th>Reach</th>
+                        <th>Cruise</th>
+                        <th>Takeoff</th>
+                        <th>Hourly</th>
+                    </tr>
+                    </thead>
+                    <tbody>
+                        <td>{id}</td>
+                        <td>{planeName}</td>
+                        <td>{manufacturer}</td>
+                        <td>{model}</td>
+                        <td>{price}</td>
+                        <td>{seats}</td>
+                        <td>{reach}</td>
+                        <td>{cruise}</td>
+                        <td>{takeoff}</td>
+                        <td>{hourly}</td>
+                    </tbody>
+            <table/>
+                <div class="text-center">
+                <a href="/fleet">Go back</a>
+                </div>
+        </div>
+
+    </div>
+</div>