+++ changelog.md
@@ -1,3 +1,42 @@
 #Change Log Team membership: Matthew Li (Captain), Hansol Lee (Mate), Vincent Lee (Mate), Clinton Bock (Mate)
 Team conventions: Allman notation, markdown for changelog
 Changelog format: Markdown
+
+## Clint Bock - 2017-10-05 2:00 PM 
+- Added the Airplane Model to the models folder
+
+## Hansol Lee - 2017-10-05 2:00PM
+- Added Airports model
+
+## Vincent Lee - 2017-10-05 3:11 PM
+- Set base template
+- Added external stylesheets and scripts
+
+## Matthew Li - 2017-10-05 4:17PM
+- Added Info controller
+- Fixed naming conventions in models
+
+## Vincent Lee - 2017-10-05 11:00 PM
+- Added flight, airport, airplane info on dashboard home
+- Updated flights to show departure and arrival airport
+
+## Clint Bock - 2017-10-06 9:36 AM
+- Added the Flights Model and mock data 
+
+
+## Hansol Lee - 2017-10-06 12:31 PM
+- Added controller for fleet
+- Added controller for fleetdetail
+- Added fleet view page
+- Added fleet detail page
+
+## Vincent Lee - 2017-10-06 12:55 PM 
+- Made fixes to use codeigniter formatting
+- Made Flight Schedule page
+
+## Matthew Li 2017-10-06 9:30 PM
+- Added Info page
+- Updated side bar links
+
+## Matthew Li 2017-10-08 9:18 AM
+- Added an error page for a plane id that doesn't exist